requirements.txt
flask
flask-script
flask-sqlalchemy
sqlalchemy-migrate
numpy
WTForms